SCC: codes and characters refactoring
